py/requirements.txt
numpy==1.23.0
esdk-obs-python==3.22.2